+++ branches/v0.70/include/tmv/TMV_BandSymBandArith.h
@@ -56,10 +56,10 @@ namespace tmv {
             TMVAssert(m1.colsize() == m2.size()); 
             TMVAssert(m1.rowsize() == m2.size()); 
         }
-        inline int colsize() const { return m2.size(); }
-        inline int rowsize() const { return m2.size(); }
-        inline int nlo() const { return TMV_MAX(m1.nlo(),m2.nlo()); }
-        inline int nhi() const { return TMV_MAX(m1.nhi(),m2.nhi()); }
+        inline ptrdiff_t colsize() const { return m2.size(); }
+        inline ptrdiff_t rowsize() const { return m2.size(); }
+        inline ptrdiff_t nlo() const { return TMV_MAX(m1.nlo(),m2.nlo()); }
+        inline ptrdiff_t nhi() const { return TMV_MAX(m1.nhi(),m2.nhi()); }
         inline T getX1() const { return x1; }
         inline const GenBandMatrix<T1>& getM1() const { return m1; }
         inline T getX2() const { return x2; }
@@ -249,11 +249,11 @@ namespace tmv {
             const GenSymBandMatrix<T2>& _m2) :
             x(_x), m1(_m1), m2(_m2)
         { TMVAssert(m1.rowsize() == m2.colsize()); }
-        inline int colsize() const { return m1.colsize(); }
-        inline int rowsize() const { return m2.rowsize(); }
-        inline int nlo() const 
+        inline ptrdiff_t colsize() const { return m1.colsize(); }
+        inline ptrdiff_t rowsize() const { return m2.rowsize(); }
+        inline ptrdiff_t nlo() const 
         { return TMV_MIN(colsize()-1,m1.nlo()+m2.nlo()); }
-        inline int nhi() const 
+        inline ptrdiff_t nhi() const 
         { return TMV_MIN(rowsize()-1,m1.nhi()+m2.nhi()); }
         inline T getX() const { return x; }
         inline const GenBandMatrix<T1>& getM1() const { return m1; }
@@ -293,11 +293,11 @@ namespace tmv {
             const GenBandMatrix<T2>& _m2) :
             x(_x), m1(_m1), m2(_m2)
         { TMVAssert(m1.rowsize() == m2.colsize()); }
-        inline int colsize() const { return m1.colsize(); }
-        inline int rowsize() const { return m2.rowsize(); }
-        inline int nlo() const 
+        inline ptrdiff_t colsize() const { return m1.colsize(); }
+        inline ptrdiff_t rowsize() const { return m2.rowsize(); }
+        inline ptrdiff_t nlo() const 
         { return TMV_MIN(colsize()-1,m1.nlo()+m2.nlo()); }
-        inline int nhi() const
+        inline ptrdiff_t nhi() const
         { return TMV_MIN(rowsize()-1,m1.nhi()+m2.nhi()); }
         inline T getX() const { return x; }
         inline const GenSymBandMatrix<T1>& getM1() const { return m1; }
