+++ branches/v0.70/include/tmv/TMV_BaseSymMatrix.h
@@ -76,7 +76,7 @@ namespace tmv {
     {
         typedef TMV_RealType(T) RT;
         typedef TMV_ComplexType(T) CT;
-        virtual int size() const = 0;
+        virtual ptrdiff_t size() const = 0;
         virtual SymType sym() const = 0;
         inline bool issym() const { return isReal(T()) || (sym() == Sym); }
         inline bool isherm() const
